pyproject.toml
[tool.poetry]
name = "remote-code"
version = "0.0.1"
description = "Create Self-Hosted CDEs on AWS in Seconds"
authors = ["Harsh Mishra <hmisraji07@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
isort = "^5.13.2"
ansible = "^9.2.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rcode = "remote_code.cli.__main__:cli"